src/main/java/cinema/service/impl/ShoppingCartServiceImpl.java
package cinema.service.impl;

import cinema.dao.ShoppingCartDao;
import cinema.dao.TicketDao;
import cinema.model.MovieSession;
import cinema.model.ShoppingCart;
import cinema.model.Ticket;
import cinema.model.User;
import cinema.service.ShoppingCartService;
import org.springframework.stereotype.Service;

@Service
public class ShoppingCartServiceImpl implements ShoppingCartService {
    private final ShoppingCartDao shoppingCartDao;
    private final TicketDao ticketDao;

    public ShoppingCartServiceImpl(ShoppingCartDao shoppingCartDao, TicketDao ticketDao) {
        this.shoppingCartDao = shoppingCartDao;
        this.ticketDao = ticketDao;
    }

    @Override
    public void addSession(MovieSession movieSession, User user) {
        if (user == null || movieSession == null
                || user.getId() == null || movieSession.getId() == null) {
            throw new RuntimeException("There's no such user or movie session: "
                    + user + ", " + movieSession);
        }
        Ticket ticket = new Ticket();
        ticket.setMovieSession(movieSession);
        ticket.setUser(user);
        ShoppingCart shoppingCart = shoppingCartDao.getByUser(user);
        ticketDao.add(ticket);
        shoppingCart.getTickets().add(ticket);
        shoppingCartDao.update(shoppingCart);
    }

    @Override
    public ShoppingCart getByUser(User user) {
        if (user == null || user.getId() == null) {
            throw new RuntimeException("There's no such user " + user);
        }
        return shoppingCartDao.getByUser(user);
    }

    @Override
    public void registerNewShoppingCart(User user) {
        if (user == null || user.getId() == null) {
            throw new RuntimeException("There's no such user " + user);
        }
        ShoppingCart shoppingCart = new ShoppingCart();
        shoppingCart.setUser(user);
        shoppingCartDao.add(shoppingCart);
    }

    @Override
    public void clear(ShoppingCart shoppingCart) {
        if (shoppingCart == null || shoppingCart.getId() == null) {
            throw new RuntimeException("There's no such shopping cart " + shoppingCart);
        }
        shoppingCart.setTickets(null);
        shoppingCartDao.update(shoppingCart);
    }
}
